# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Employees API",
        "termsOfService": "http://somewhere.com/",
        "contact": {
            "name": "API Support",
            "url": "http://somewhere.com/support",
            "email": "support@somewhere.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/auth/changePassword": {
            "put": {
                "description": "ChangePassword",
                "consumes": [
                    "application/json"
                ],
                "summary": "ChangePassword",
                "parameters": [
                    {
                        "description": "ChangePassword Data",
                        "name": "changePassword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.AuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/auth.AuthResponse2"
                        }
                    },
                    "400": {
                        "description": "BadRequest",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "UnAuthorize",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "NotFound",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "InternalServerError",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/auth/currentUser": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "GetCurrentUser",
                "consumes": [
                    "application/json"
                ],
                "summary": "Get CurrentUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/auth.EmployeeResponse"
                        }
                    },
                    "400": {
                        "description": "BadRequest",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "UnAuthorize",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "NotFound",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "InternalServerError",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "description": "Login to System",
                "consumes": [
                    "application/json"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login Data",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.AuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/auth.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "BadRequest",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "UnAuthorize",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "NotFound",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "InternalServerError",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/employees": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Getting all employee from database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "get all employee",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search by firstName",
                        "name": "firstName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search by lastName",
                        "name": "lastName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search by role",
                        "name": "role",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/employees.ListResponse"
                        }
                    },
                    "401": {
                        "description": "UnAuthorize",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "NotFound",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "InternalServerError",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Adding new employee to database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "create employee",
                "parameters": [
                    {
                        "description": "Employee Data",
                        "name": "employee",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/employees.EmployeeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/employees.ListResponse"
                        }
                    },
                    "400": {
                        "description": "BadRequest",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "UnAuthorize",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "NotFound",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "InternalServerError",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/employees/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "getting employee by id from database",
                "consumes": [
                    "application/json"
                ],
                "summary": "get employee by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/employees.ListResponse"
                        }
                    },
                    "400": {
                        "description": "BadRequest",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "UnAuthorize",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "NotFound",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "InternalServerError",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "update employee to database",
                "consumes": [
                    "application/json"
                ],
                "summary": "update employee",
                "parameters": [
                    {
                        "description": "Employee Data",
                        "name": "employee",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/employees.EmployeeRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/employees.ListResponse"
                        }
                    },
                    "400": {
                        "description": "BadRequest",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "UnAuthorize",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "NotFound",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "InternalServerError",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "delete employee from database",
                "consumes": [
                    "application/json"
                ],
                "summary": "delete employee",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "BadRequest",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "UnAuthorize",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "NotFound",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "InternalServerError",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.AuthRequest": {
            "type": "object",
            "properties": {
                "employeeId": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "auth.AuthResponse": {
            "type": "object",
            "properties": {
                "employeeId": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "auth.AuthResponse2": {
            "type": "object",
            "properties": {
                "employeeId": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "auth.EmployeeResponse": {
            "type": "object",
            "properties": {
                "employeeId": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "salary": {
                    "type": "number"
                }
            }
        },
        "employees.EmployeeRequest": {
            "type": "object",
            "properties": {
                "employeeId": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "salary": {
                    "type": "number"
                }
            }
        },
        "employees.EmployeeResponse": {
            "type": "object",
            "properties": {
                "employeeId": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "salary": {
                    "type": "number"
                }
            }
        },
        "employees.ListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/employees.EmployeeResponse"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}